Refactor http send file tests to also run with SSL
- new tests for SSL reveal problem with sendFileAsync4096/sendFileAsync1024
  when SSL is used
